Remove duplicate classes
