--- src/main/java/org/team1540/robot2019/commands/climber/Level3Climb.java
@@ -1,12 +1,0 @@
-package org.team1540.robot2019.commands.climber;
-
-import edu.wpi.first.wpilibj.command.CommandGroup;
-
-public class Level3Climb extends CommandGroup {
-
-    public Level3Climb() {
-        addSequential(new PrepareForClimb());
-        addSequential(new ClimbLevelThree());
-    }
-
-}
